Move publisher CLI from tools/ to cmd/
Move the publisher CLI from tools/publisher/ to cmd/publisher/ as it's a proper application rather than a development utility. Update all references including:
- Makefile build and clean targets
- README.md documentation links
- VS Code launch configuration
- Integration test script
- FAQ documentation
- Go import paths

Addresses feedback from PR #324.
