+++ cmd/publisher/README.md
