+++ cmd/publisher/auth/common.go
