+++ cmd/publisher/auth/dns.go
