+++ cmd/publisher/auth/http.go
