+++ cmd/publisher/auth/interface.go
