+++ cmd/publisher/auth/none.go
