+++ cmd/publisher/build.sh
