+++ cmd/publisher/commands/init.go
