+++ cmd/publisher/commands/logout.go
