+++ cmd/publisher/commands/publish.go
